% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlc.R
\name{align_layers}
\alias{align_layers}
\title{Harmonise projection, resolution and extent of a list of rasters}
\usage{
align_layers(x)
}
\arguments{
\item{x}{list of RasterLayer objects}
}
\value{
Another list of RasterLayer objects
}
\description{
Convert a list of rasters to a common projection, resolution and extent.
}
\details{
The final projection is the most common one, other than NA. If a raster has
unknown projection, it is assumed to be equal to the most common one.

All rasters are resampled to the same extent and resolution as the \emph{first}
raster in the list.
}
