% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borders0.R
\name{borders0}
\alias{borders0}
\title{Outer boundaries of a vector map}
\usage{
borders0(x)
}
\arguments{
\item{x}{A SpatialPolygons* object}
}
\value{
A single polygon with the outermost boundary.
}
\description{
Outer boundaries of a vector map
}
\examples{
  cmr <- mapMCDA_datasets()
  sp::plot(cmr$cmr_admin3)
  sp::plot(borders0(cmr$cmr_admin3))
}
