% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_sep.R
\name{csv_sep}
\alias{csv_sep}
\title{Detection of separator in csv files}
\usage{
csv_sep(x)
}
\arguments{
\item{x}{character. File name.}
}
\value{
Character representation of the field separator.
}
\description{
Returns the separator that consistently separates the lines in the same
number of fields, or fails.
}
\details{
Tested separators are: ",", " ", "\\t", "|", ";", ":"
}
\examples{
  tf <- tempfile()
  td <- data.frame(T = LETTERS[1:10], N = 1:10)
  write.csv(td, tf, row.names = FALSE)
  csv_sep(tf)  # ","
  write.csv2(td, tf, row.names = FALSE)
  csv_sep(tf)  # ";"
}
