% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance_map}
\alias{distance_map}
\title{Compute a distance map to spatial features}
\usage{
distance_map(x, boundaries, res = resolution(boundaries, min_ncells =
  100))
}
\arguments{
\item{x}{Spatial* object with relevant spatial features.}

\item{boundaries}{SpatialPolygon* object with administrative borders.}

\item{res}{numeric. Resolution for the outcome distance map. The default divides
the smallest dimension into 100 cells.}
}
\value{
A raster map with distances in meters to the closest spatial feature.
}
\description{
Returns a distance map (in meters) to spatial features.
}
\details{
Distances are computed up to the given boundaries and at given resolution.
}
\examples{
  cmr <- mapMCDA_datasets()
  raster::plot(distance_map(cmr$water_bodies, boundaries = cmr$cmr_admin3))
}
