% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance_to_vector}
\alias{distance_to_vector}
\title{Distance to vector features}
\usage{
distance_to_vector(x, mask)
}
\arguments{
\item{x}{Spatial* object}

\item{mask}{Raster* object}
}
\value{
Raster* object. Distance map in meters.
}
\description{
Compute a raster file with the distances to the nearest vector feature,
within the limits of a given mask.
}
\details{
The resolution of the mask determines that of the outcome.
}
\examples{
  cmr <- mapMCDA_datasets()
  dist_wb <- distance_to_vector(cmr$water_bodies, cmr$animal.density)
  raster::plot(dist_wb)
}
