% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_country_admin.R
\name{download_country_admin}
\alias{download_country_admin}
\title{Download Country Administrative Borders from GADM}
\usage{
download_country_admin(country_name, level = 3, path = tempdir())
}
\arguments{
\item{country_name}{character. A valid country name.}

\item{level}{integer 0:3.}

\item{path}{character. Where to store the downloaded object.}
}
\value{
A SpatioalPolygonsDataFrame object.
}
\description{
Download Country Administrative Borders from GADM
}
\examples{
  cmr <- download_country_admin("Cameroon")
}
