% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{epidemic_threshold}
\alias{epidemic_threshold}
\title{Compute the Epidemic Threshold of a graph}
\usage{
epidemic_threshold(x, beta = 1)
}
\arguments{
\item{x}{an \code{igraph} object}

\item{beta}{numeric, between 0 and 1. Probability of transmission.}
}
\value{
a list the weighted and unweighted Potential for
  Transmission \eqn{R_0}{R₀} and its inverse, the Epidemic
  Threshold \eqn{q}. As an attribute named "sna", a data.frame with
  the in/out-degrees of each node and their individual contribution
  to R0.

If the input graph is unweighted, the weighted component is NULL.
}
\description{
Weighted and unweighted Epidemic Threshold \eqn{q} of a graph.
}
\details{
The Epidemic Threshold \eqn{q} quantifies the minimal expeced
transmission coefficient necessary for diffusing an epidemy in a
network. It is computed as the inverse of the \emph{Potential for
transmission} of the network: a measure of the expected number of
nodes affected by an infectious node, which is a generalisation of
the Basic Reproduction Number \eqn{R_0}{R₀} of an epidemy to the
context of a network. It thus quantifies the potential for
transmission of an infection throughout the contact network. It is
computed in terms of the incoming-outgoing rates from the network's
nodes: \deqn{R_0 = \beta \frac{\hat{k_\mathrm{in}
k_\mathrm{out}}}{\hat{k_\mathrm{in}}},}{R₀ = \beta〈k_in*k_out〉/〈k_in〉,}
where \eqn{\beta} is the transmission coefficient among animals,
\eqn{k_\mathrm{in/out}}{k_in/out} are the in/out-degrees of a node
and the \eqn{\hat{\cdot}}{〈·〉} symbol represents the average value
across all nodes in the graph.

The unweighted value computed above is most appropriate for a
highly infectious epidemy with high animal-prevalence on nodes, as
it assumes that any contact is potentially infectious.

In the weighted formulation, \eqn{k_\mathrm{in/out}}{k_in/out} are
the weight values for the incoming/outgoing edges in each node. It
is more appropriate for low-prevalence diseases, where the
transmission probability is assumed proportional to the number of
contacts.

The default value of 1 for the probability of transmission
\code{beta} implies that every infectious contact leads to
transmission.
}
\examples{
  g <- igraph::graph_from_literal(A --+ B --+C, A --+C, B --+D)
  epidemic_threshold(g)
  
  ## weighted graph
  igraph::E(g)$weight <- c(10, 1, 2, 5)
  epidemic_threshold(g)
}
\references{
Volkova VV, Howey R, Savill NJ, Woolhouse MEJ (2010) Sheep
  Movement Networks and the Transmission of Infectious Diseases.
  PLoS ONE 5(6): e11185.
  https://doi.org/10.1371/journal.pone.0011185
}
