% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.reciprocal.R
\name{is.reciprocal}
\alias{is.reciprocal}
\title{Check reciprocal matrix}
\usage{
is.reciprocal(x)
}
\arguments{
\item{x}{square numeric matrix or number}
}
\value{
logical
}
\description{
Check whether a matrix is reciprocal. This is, whether the
entrywise product of itself with its transpose (Hadamard product)
is 1 for all its elements.
}
\examples{
  Xf <- matrix(1:16, 4, 4)
  Xr <- Xf
  Xr[upper.tri(Xf)] <- 1/t(Xf)[upper.tri(Xf)]
  diag(Xr) <- 1
  mapMCDA:::is.reciprocal(Xf)  # FALSE
  mapMCDA:::is.reciprocal(Xr)  # TRUE
}
