% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_type.R
\name{layer_type}
\alias{layer_type}
\title{Detect type of geographic information}
\usage{
layer_type(x)
}
\arguments{
\item{x}{string. A file name.}
}
\value{
character. Either "vector", "raster" or "network". NA for unknown formats.
}
\description{
Either "vector", "raster" or "network" depending on the file extension.
}
\examples{
  layer_type("myfile.tif")
}
