% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_layer.R
\name{load_dir}
\alias{load_dir}
\title{Load all layers within a directory}
\usage{
load_dir(path)
}
\arguments{
\item{path}{String. The path to the directory where spatial layers are
stored.}
}
\value{
list of layers in the form of Spatial* or RasterLayer objects
}
\description{
Load all recognised vector and raster files within a given directory.
}
\examples{
   pkg_layers <- system.file("cartography/CMR/", package = "mapMCDA")
   cmr <- load_dir(pkg_layers)
}
