% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_layer.R
\name{load_layer}
\alias{load_layer}
\title{Load a map layer}
\usage{
load_layer(x)
}
\arguments{
\item{x}{File name with full path}
}
\value{
A \code{Spatial*} object for vector maps and a \code{RasterLayer}
  object for raster maps.
}
\description{
Automatically detects whether it is a vector or raster map and uses the
appropriate read function.
}
\examples{
 (pkg_layers <- list.files(
   system.file("cartography/CMR", package = "mapMCDA"),
   full.names = TRUE
 ))
 
 epiunits <- load_layer(pkg_layers[[2]])
}
