% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMCDA_app.R
\name{mapMCDA_app}
\alias{mapMCDA_app}
\title{mapMCDA Shiny web interface}
\usage{
mapMCDA_app(bg = FALSE)
}
\arguments{
\item{bg}{Logical. If TRUE, the GUI is launched in the background, leaving
available the current session.}
}
\description{
Launches the graphical interface (GUI) of the package.
}
