% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_deps.R
\name{pkg_deps}
\alias{pkg_deps}
\title{mapMCDA package dependencies}
\usage{
pkg_deps()
}
\value{
character vector of package names
}
\description{
The list of packages has been recovered after the installation of the package
on a clean Windows 7 virtual machine.
}
