% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\docType{methods}
\name{rasterize,igraph,Raster-method}
\alias{rasterize,igraph,Raster-method}
\title{Rasterize a geographic network}
\usage{
\S4method{rasterize}{igraph,Raster}(x, y, field = "importance", ...)
}
\description{
Copmute a raster with the importance of the nearest network node.
}
\details{
This assumes that the network object has node attributes "Lon" "Lat"
in the WGS84 reference system.
}
