% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\docType{methods}
\name{rasterize_geonetwork}
\alias{rasterize_geonetwork}
\alias{rasterize,geonetwork,Raster-method}
\alias{rasterize,geonetwork,SpatialPolygons-method}
\title{Rasterize a geographic network}
\usage{
\S4method{rasterize}{geonetwork,Raster}(x, y, field = "importance", ...)

\S4method{rasterize}{geonetwork,SpatialPolygons}(x, y,
  res = resolution(y, min_ncells = 100), ...)
}
\arguments{
\item{x}{a geographic network (a geonetwork object with node attributes
"Lon" and "Lat" in the WGS84 reference system)}

\item{y}{a Raster* or a SpatialPolygons* object.}

\item{field}{character. The attribute of the network nodes to be
assigned to the voronoi polygon associated with each node. By
default it computes the \emph{importance} of each node in the
network. But you can manually compute any other measure of
interest, and pass it as a vertex attribute (see
\code{?igraph::`vertex_attr<-`}).}

\item{...}{Other arguments passed on to
\code{\link[raster]{rasterize}}.}

\item{res}{numeric vector of length 2. Resolution in Lon/Lat of
rasterisation.}
}
\description{
Copmute a raster with the importance of the nearest network node.
}
\details{
This assumes that the network object has node attributes "Lon" "Lat"
in the WGS84 reference system.

If the graph is weighted, the weighted importance will be computed.
If you want the unweighted importance, remove the weight attribute
from the graph with \code{igraph::delete_edge_attr(x, "weight")}.

The importance of each node is computed as the relative
contribution to the Potential for Transmision \eqn{R_0} of the
graph. I.e., \eqn{R_{0k}/R_0}. See \code{\link{epidemic_threshold}}.
}
