% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{read_network}
\alias{read_network}
\title{Read and interpret network data}
\usage{
read_network(x)
}
\arguments{
\item{x}{character. Path of csv file}
}
\value{
object of class \code{igraph}
}
\description{
Read a csv file with a specific format (see Details) and intepret
it as network directed data with possibly weighted edges.
}
\details{
The file must be plain text with comma-separated columns and variable
names in the first line.
Field separators other than comma are also supported. In particular,
semi-colon field separators and comma decimal separator.
There must be either 6 or 7 columns in the same order and of the
same types as follows:
\itemize{
  \item origin       character
  \item destination  character
  \item lon_orig     numeric (decimal degrees, WGS84)
  \item lat_orig     numeric (decimal degrees, WGS84)
  \item lon_dest     numeric (decimal degrees, WGS84)
  \item lat_dest     numeric (decimal degrees, WGS84)
  \item volume       Optional. directed flux in some consistent unit
}

Variable names can be different. If strings contain spaces they must
be quoted.
}
\examples{
  d <- data.frame(from = "A", to = "B", fx = 0, fy = 0, tx = 1, ty = 1)
  tf <- tempfile()
  write.csv(d, tf, row.names = FALSE)
  read_network(tf)
}
