% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{resolution}
\alias{resolution}
\title{Compute a suitable resolution for a map}
\usage{
resolution(x, min_ncells = 100)
}
\arguments{
\item{x}{a Raster* or Spatial* object}

\item{min_ncells}{Number of cells for the smallest dimension}
}
\value{
A number. The resolution that respects the desired number of cells.
}
\description{
Compute a resolution in the original map units that conforms to given
criteria.
}
\examples{
  r <- raster::raster(nrows = 5, ncols = 11, xmn = 0, xmx = 11, ymn = 0, ymx = 5)
  resolution(r, min_ncells = 10)
}
