% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_layer.R
\name{risk_layer}
\alias{risk_layer}
\title{Compute risk layer}
\usage{
risk_layer(x, boundaries, scale_target = c(0, 100))
}
\arguments{
\item{x}{a Spatial*, RasterLayer or igraph object}

\item{boundaries}{a Spatial* object, used to determine the boundaries of the
computed risk layer.}

\item{scale_target}{numeric vector of length 2. New scale.}
}
\value{
A RasterLayer object in the new scale.
}
\description{
Compute a raster map from the input layer and rescale with a linear
relationship.
}
\details{
For Spatial* objects (geometries such as point, lines or polygons),
compute the \code{distance_map()}, which gives a RasterLayer. For
\code{igraph} objects (from network data), compute a RasterLayer 
with the relative importance of the nearest node. For a RasterLayer
mask, extend or crop to the \code{boundaries} as needed.

Finally, scale the RasterLayer outcome of any of the three input
types. If you need an inverse relationship, just reverse the target
scale.
}
\examples{
  ad <- mapMCDA_datasets()$animal.density
  bd <- mapMCDA_datasets()$cmr_admin3
  raster::plot(ad)
  raster::plot(risk_layer(ad, bd, scale_target = c(-1, 1)))
  raster::plot(risk_layer(ad, bd, scale_target = c(1, -1)))
}
