% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_unit.R
\name{risk_plot}
\alias{risk_plot}
\title{Produce final plot of risk}
\usage{
risk_plot(x, v, n)
}
\arguments{
\item{x}{SpatialPolygons*. Epidemiological units.}

\item{v}{Numeric. Risk values.}

\item{n}{Number of risk levels.}
}
\value{
spplot
}
\description{
Represent the value of risk by epidemiological unit
categorised into n risk-levels
}
\examples{
  epi_units <- mapMCDA_datasets()$cmr_admin3
  risk_plot(epi_units, seq_along(epi_units), n = 4)
}
