% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_unit.R
\name{risk_table}
\alias{risk_table}
\title{Produce final table of risk levels by epidemiological unit}
\usage{
risk_table(x, v, n)
}
\arguments{
\item{x}{SpatialPolygons*. Epidemiological units.}

\item{v}{Numeric. Risk values.}

\item{n}{Number of risk levels.}
}
\value{
data.frame
}
\description{
Represent the value of risk by epidemiological unit
categorised into n risk-levels
}
\examples{
  epi_units <- mapMCDA_datasets()$cmr_admin3
  risk <- runif(nrow(epi_units), 0, 100)
  risk_table(epi_units, risk, n = 4)
}
