% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_unit.R
\name{risk_unit}
\alias{risk_unit}
\title{Risk by epidemiological unit}
\usage{
risk_unit(r, eu, fun = mean)
}
\arguments{
\item{r}{RasterLayer object}

\item{eu}{SpatialPolygons* object}

\item{fun}{A summary function. By default the mean.}
}
\value{
numeric vector of the same length as the number of epidemiological
  units.
}
\description{
Compute the risk for each of the epidemiological units.
}
\details{
Summarises the level of risk using a function such as the mean or median
value across the epidemiological unit.
}
\examples{
  cmr <- mapMCDA_datasets()
  cmr$cmr_admin3$rv <- risk_unit(cmr$animal.density, cmr$cmr_admin3)
  spplot(cmr$cmr_admin3[, "rv"], cuts = 3)
}
