% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmext.R
\name{rmext}
\alias{rmext}
\title{Remove filename extension}
\usage{
rmext(x)
}
\arguments{
\item{x}{Character}
}
\description{
Removes the trailing characters after and including the last point in a
string.
}
\examples{
  rmext("some/long.filename.extension")
}
