% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{voronoi}
\alias{voronoi}
\title{Voronoi polygons}
\usage{
voronoi(x, ext, eps = 1e-09)
}
\arguments{
\item{x}{SpatialPoints* or two-column matrix or data.frame with x and y coordinates.}

\item{ext}{Extent object.}

\item{eps}{Numerical tolerance used in triangulation. See \code{\link[deldir]{deldir}}.}
}
\value{
SpatialPolygonsDataFrame.
}
\description{
Compute a vornoi tesselation of a map for a set of points.
}
\author{
Borrowed from function \code{\link[dismo]{voronoi}} from
  package \code{dismo} by R. Hijmans et al., which ultimately
  uses the \code{\link[deldir]{deldir}} function by Rolf Turner.
}
