% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlc.R
\name{wlc}
\alias{wlc}
\title{Weighted Linear Combination of risk factors}
\usage{
wlc(x, w)
}
\arguments{
\item{x}{list of risk factors (RaterLayer objects)}

\item{w}{vector of weights}
}
\value{
A RasterLayer object.
}
\description{
Weighted Linear Combination of risk factors
}
