% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_layer.R
\name{risk_layer}
\alias{risk_layer}
\title{Compute risk layer}
\usage{
risk_layer(x, boundaries, scale_target = c(0, 100))
}
\arguments{
\item{x}{a RasterLayer object}

\item{boundaries}{a Spatial* object, used to determine the boundaries of the
computed risk layer.}

\item{scale_target}{numeric vector of length 2. New scale.}
}
\value{
A RasterLayer object in the new scale.
}
\description{
Rescale a raster with a linear relationship.
}
\details{
If you need an inverse relationship, just reverse the target scale.
}
\examples{
  ad <- mapMCDA_datasets()$animal.density
  bd <- mapMCDA_datasets()$cmr_admin3
  raster::plot(ad)
  raster::plot(risk_layer(ad, bd, scale_target = c(-1, 1)))
  raster::plot(risk_layer(ad, bd, scale_target = c(1, -1)))
}
